feat : Party Backend Refactoring
